requirements.txt
numpy~=1.24.2
torch~=1.13.1
torchaudio~=0.13.1
matplotlib~=3.6.3
seaborn~=0.12.2
pandas~=1.5.3
PyYAML~=6.0
tqdm~=4.64.1
scikit-learn~=1.2.1
speechbrain~=0.5.13
scipy~=1.10.0